requirements.txt
absl-py==2.1.0
alabaster==0.7.16
astroid==3.1.0
astunparse==1.6.3
Babel==2.14.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
dm-tree==0.1.8
docutils==0.20.1
flatbuffers==24.3.7
fonttools==4.49.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
imagesize==1.4.1
importlib_metadata==7.0.2
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.3.1
Jinja2==3.1.3
joblib==1.3.2
keras==3.0.5
keyring==24.3.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mypy-extensions==1.0.0
namex==0.0.7
nh3==0.2.15
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.3
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.2
pytest==8.1.1
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
readme_renderer==43.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
setuptools==69.1.1
SimpleWebSocketServer==0.1.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
textblob==0.18.0.post0
tomlkit==0.12.4
tqdm==4.66.2
twine==5.0.0
typing==3.7.4.3
typing_extensions==4.10.0
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
zipp==3.18.0
